+++ tests/models/test_attribute_decorators.py
@@ -79,3 +79,56 @@ def test_on_deletion_parenthesis(sheraf_connection):
     assert not Model.deleted
     m.delete()
     assert Model.deleted
+
+
+def test_initialization_order_on_creation(sheraf_connection):
+    class Model(tests.IntAutoModel):
+        foo = sheraf.SimpleAttribute(default="before_foo")
+        bar = sheraf.SimpleAttribute(default="before_bar")
+
+        @foo.on_creation
+        @bar.on_creation
+        def change(self, new=None, old=None):
+            assert self.foo == "before_foo"
+            assert self.bar == "before_bar"
+            yield
+            assert self.foo == "after_foo"
+            assert self.bar == "after_bar"
+
+    Model.create(foo="after_foo", bar="after_bar")
+
+
+def test_initialization_order_on_update(sheraf_connection):
+    class Model(tests.IntAutoModel):
+        foo = sheraf.SimpleAttribute(default="before_foo")
+        bar = sheraf.SimpleAttribute(default="before_bar")
+
+        @foo.on_edition
+        @bar.on_edition
+        def change(self, new=None, old=None):
+            assert self.foo == "before_foo"
+            assert self.bar == "before_bar"
+            yield
+            assert self.foo == "after_foo"
+            assert self.bar == "after_bar"
+
+    m = Model.create()
+    m.update(foo="after_foo", bar="after_bar")
+
+
+def test_initialization_order_on_deletion(sheraf_connection):
+    class Model(tests.IntAutoModel):
+        foo = sheraf.SimpleAttribute(default="before")
+        bar = sheraf.SimpleAttribute(default="before")
+
+        @foo.on_deletion
+        @bar.on_deletion
+        def change(self, new=None, old=None):
+            assert self.foo == "before"
+            assert self.bar == "before"
+            yield
+            assert not self.attributes["foo"].is_created(self)
+            assert not self.attributes["bar"].is_created(self)
+
+    m = Model.create()
+    m.delete()